moved shading geometryFilter.ts into main package. Filter gets now applied in scene.calculate as well
